requirements/base.txt
Django==1.9.1
psycopg2==2.5.4
django-leaflet
gunicorn==19.3.0
requests[security]
xlrd
djangorestframework==3.3.2
djangorestframework-csv==1.4.1
django-rest-swagger==0.3.5
django-cors-headers==1.1.0
openpyxl==2.3.3
numpy==1.10.4
pandas==0.19.1
scipy==0.17.0
django-allauth==0.25.2
lxml==3.6.4
premailer
django-anymail==0.6.1
django-crispy-forms
requests-futures
google-api-python-client==1.5.3
pip==8.1.2
setuptools==27.3.0
ebmdatalab-python==0.0.14
matplotlib==1.5.3
